keep Function shr impl in mod
